cast to float first for cast

+++ ntt/include/nncase/ntt/primitive_ops.h
@@ -334,7 +334,7 @@ template <class T1, class T2> struct clamp {
 
 template <class T1, class T2> struct cast {
     constexpr T2 operator()(const T1 &v) const noexcept {
-        return static_cast<T2>(v);
+        return static_cast<T2>(static_cast<float>(v));
     }
 };
